Refactor: 댓글 입력창 컴포넌트화로 불필요한 리랜더링 개선 및 게시버튼 onSubmit시 비활성화
