+++ src/pages/leeseyoon/Main/components/Article/components/NewComment.scss
@@ -1,0 +1,147 @@
+article {
+  margin-bottom: 12px;
+  border: 1px solid #dbdbdb;
+  border-radius: 5px;
+  background-color: #ffffff;
+
+  .feed-header {
+    display: flex;
+    justify-content: space-between;
+    padding: 16px;
+
+    .left {
+      display: flex;
+      align-items: center;
+
+      .new-recommended-img-box {
+        width: 50px;
+        height: 50px;
+        /* border: 2.5px solid #dbdbdb; */
+        margin-right: 12px;
+        overflow: hidden;
+        border-radius: 50%;
+        /* background: radial-gradient(rgba(0, 0, 0, 0.15) 60%, transparent 0),
+              radial-gradient(white 62%, transparent 0),
+              linear-gradient(to top right, orange, deeppink); */
+      }
+    }
+
+    .right {
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+    }
+  }
+
+  .feed-img {
+    overflow: hidden;
+
+    img {
+      width: 100%;
+      max-height: 550px;
+      object-fit: cover;
+    }
+  }
+
+  .section-wrap {
+    padding: 5px 16px 16px 16px;
+
+    .feed-body {
+      display: flex;
+      justify-content: space-between;
+      margin-bottom: 5px;
+
+      .left {
+        display: flex;
+        justify-content: space-between;
+        /* background-color: #0095f6; */
+        width: 85px;
+      }
+
+      i {
+        font-size: larger;
+      }
+    }
+
+    .like-wrap {
+      display: flex;
+      align-items: center;
+      padding: 0px 0px 5px 0px;
+
+      .like-poeple-img-box {
+        width: 28px;
+        height: 28px;
+        /* border: 2.5px solid #dbdbdb; */
+        margin-right: 5px;
+        overflow: hidden;
+        border-radius: 50%;
+        /* background: radial-gradient(rgba(0, 0, 0, 0.15) 60%, transparent 0),
+              radial-gradient(white 62%, transparent 0),
+              linear-gradient(to top right, orange, deeppink); */
+      }
+    }
+
+    .new-comment {
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+      justify-content: space-between;
+      margin-top: 10px;
+      padding: 10px 0px 0px 0px;
+      border-top: 1px solid #dbdbdb;
+
+      i {
+        font-size: 19px;
+      }
+
+      input {
+        width: 440px;
+        border: 0px;
+      }
+
+      button {
+        border: 0px;
+        background-color: #ffffff;
+        color: #0095f6;
+        cursor: pointer;
+
+        &:disabled {
+          opacity: 0.3;
+          cursor: default;
+        }
+      }
+    }
+
+    .text {
+      margin-bottom: 5px;
+    }
+
+    .comment {
+      /* height: calc(100% - 29.5px); */
+      width: 400px;
+      overflow: hidden;
+      list-style: none;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+
+      strong {
+        font-weight: bold;
+      }
+    }
+  }
+}
+
+.font-weight-600 {
+  font-weight: 600;
+}
+
+.new-story-img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  clip-path: circle(42%);
+}
+
+.font-size-14 {
+  font-size: 15px;
+}
